+++ nkululeko/feat_extract/feats_import.py
@@ -1,10 +1,11 @@
 # feats_import.py
 
+import os
+import ast
+import audformat
+import pandas as pd
 from nkululeko.utils.util import Util
 from nkululeko.feat_extract.featureset import Featureset
-import os
-import pandas as pd
-import audformat
 
 
 class Importset(Featureset):
@@ -14,32 +15,29 @@ class Importset(Featureset):
         super().__init__(name, data_df)
 
     def extract(self):
-        """Import the features or load them from disk if present."""
-        store = self.util.get_path("store")
-        storage = f"{store}{self.name}.pkl"
-        extract = eval(self.util.config_val("FEATS", "needs_feature_extraction", False))
-        no_reuse = eval(self.util.config_val("FEATS", "no_reuse", "False"))
-        feat_import_file = self.util.config_val("FEATS", "import_file", False)
-        if not os.path.isfile(feat_import_file):
-            self.util.warn(f"no import file: {feat_import_file}")
-        if extract or no_reuse or not os.path.isfile(storage):
-            self.util.debug(f"importing features for {self.name}")
-            # df = pd.read_csv(feat_import_file, sep=',', header=0,
-            #     index_col=['file', 'start', 'end'])
+        """Import the features."""
+        self.util.debug(f"importing features for {self.name}")
+        try:
+            feat_import_files = self.util.config_val("FEATS", "import_file", False)
+            feat_import_files = ast.literal_eval(feat_import_files)
+        except ValueError as e:
+            self.util.error(
+                "feature type == import needs import_file = ['file1', 'filex']"
+            )
+        except SyntaxError as se:
+            if type(feat_import_files) == str:
+                feat_import_files = [feat_import_files]
+            else:
+                self.util.error(f"import_file is wrong: {feat_import_files}")
+
+        feat_df = pd.DataFrame()
+        for feat_import_file in feat_import_files:
+            if not os.path.isfile(feat_import_file):
+                self.util.error(f"no import file: {feat_import_file}")
             df = audformat.utils.read_csv(feat_import_file)
-            # scale features before use?
-            # from sklearn.preprocessing import StandardScaler
-            # scaler = StandardScaler()
-            # scaled_features = scaler.fit_transform(df.values)
-            # df = pd.DataFrame(scaled_features, index=df.index, columns=df.columns)
-            # use only the rows from the data index
-            # df = self.data_df.join(df).drop(columns=self.data_df.columns)
-            df = df.loc[self.data_df.index]
-            # df = pd.concat([self.data_df, df], axis=1, join="inner").drop(columns=self.data_df.columns)
-            # in any case, store to disk for later use
-            df.to_pickle(storage)
-            # and assign to be the "official" feature set
-            self.df = df
-        else:
-            self.util.debug("reusing imported features.")
-            self.df = pd.read_pickle(storage)
+            df = df[df.index.isin(self.data_df.index)]
+            feat_df = pd.concat([feat_df, df])
+        if feat_df.shape[0] == 0:
+            self.util.error(f"Imported features for data set {self.name} not found!")
+        # and assign to be the "official" feature set
+        self.df = feat_df